requirements.txt
pymongo==3.8.0
python-telegram-bot
requests==2.22.0
urllib3==1.25.2
haikunator